LightGCN-PyTorch/requirements.txt
pytorch==1.4.0
pandas==0.24.2
scipy==1.3.0
numpy==1.16.4
tensorboardX==1.8
scikit-learn==0.23.2
tqdm==4.48.2
